+++ examples/concealer_manager.rs
@@ -43,7 +43,6 @@ async fn app(window: Window, mut gfx: Graphics, mut inputs: Input) -> Result<()>
     let mut context = Context::new();
 
     //we then construct a layer. A layer is used to group and control multiple widgets together.
-    //Whenever a layer gets dropped. The widgets that belong to that layer will also be removed.
     let mut layer = context.add_layer();
 
     //lets make the first concealer. We start with a layer that it will control